Header mostly finished
Tried styling the header better, also created the "main header" on App.vue, will probably change it to a component later, but its good enough to remind me to do it later.
Still need to style better the leftside of the header, I'll learn a bit about grid layout and do it again!